# Use pure thrust