This PR converts the scan shuffle to use pure thrust and removes the second function call.

All tests passing.

Some benchmarks on V100:

| Size   | Old (G/s) | New (G/s) | Speedup |
|--------|-----------|-----------|---------|
| 2^18+1 | 3.82      | 4.42      | 1.16    |
| 2^27+1 | 3.60      | 4.07      | 1.13    |
| 2^29+1 | 3.51      | 3.965     | 1.13    |